requirements-base.txt
# Base PyTorch environment
# This file is intentionally limited to the stable PyTorch ADR workflow.

torch>=2.0
numpy>=1.23,<2
scipy>=1.10,<1.13
matplotlib>=3.6
pyyaml>=6.0
tqdm>=4.64
optuna>=3.0
